Implementing DTOs

+++ backend-api/src/main/java/com/project/backend_api/models/Genre.java
@@ -2,16 +2,13 @@ package com.project.backend_api.models;
 
 import jakarta.persistence.*;
 
-import java.util.HashSet;
-import java.util.Set;
-
 @Entity
 @Table(name = "genres")
 public class Genre {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "ID")
-    private Integer id;
+    private Long id;
 
     @Column(name = "Name", nullable = false, unique = true)
     private String name;
@@ -22,11 +19,11 @@ public class Genre {
         this.name = name;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
